ccl: move pkg/ccl/cclcli to pkg/cli
This moves the code from pkg/ccl/cclcli to pkg/cli. In the process, we
remove an unnecessary init hook.

This does produce a few more imports into ccl from pkg/cli as we have
not yet moved workloadccl or sqlproxyccl.

Epic: none
Release note: None
