+++ pkg/BUILD.bazel
@@ -40,7 +40,6 @@ ALL_TESTS = [
     "//pkg/ccl/changefeedccl/tableset:tableset_test",
     "//pkg/ccl/changefeedccl/timers:timers_test",
     "//pkg/ccl/changefeedccl:changefeedccl_test",
-    "//pkg/ccl/cliccl:cliccl_test",
     "//pkg/ccl/cloudccl/amazon:amazon_test",
     "//pkg/ccl/cloudccl/azure:azure_test",
     "//pkg/ccl/cloudccl/cloudprivilege:cloudprivilege_test",
@@ -1001,8 +1000,6 @@ GO_TARGETS = [
     "//pkg/ccl/changefeedccl/timers:timers_test",
     "//pkg/ccl/changefeedccl:changefeedccl",
     "//pkg/ccl/changefeedccl:changefeedccl_test",
-    "//pkg/ccl/cliccl:cliccl",
-    "//pkg/ccl/cliccl:cliccl_test",
     "//pkg/ccl/cloudccl/amazon:amazon_test",
     "//pkg/ccl/cloudccl/azure:azure_test",
     "//pkg/ccl/cloudccl/cloudprivilege:cloudprivilege_test",
