+++ pkg/backup/backupinfo/BUILD.bazel
@@ -80,6 +80,7 @@ go_test(
         "//pkg/blobs",
         "//pkg/ccl",
         "//pkg/cloud",
+        "//pkg/cloud/impl:cloudimpl",
         "//pkg/clusterversion",
         "//pkg/jobs/jobspb",
         "//pkg/keys",
