+++ pkg/ccl/BUILD.bazel
@@ -9,7 +9,6 @@ go_library(
         "//pkg/backup",
         "//pkg/base",
         "//pkg/ccl/changefeedccl",
-        "//pkg/ccl/cliccl",
         "//pkg/ccl/gssapiccl",
         "//pkg/ccl/jwtauthccl",
         "//pkg/ccl/ldapccl",
@@ -20,6 +19,8 @@ go_library(
         "//pkg/ccl/securityccl/fipsccl",
         "//pkg/ccl/utilccl",
         "//pkg/ccl/workloadccl",
+        "//pkg/ccl/workloadccl/cliccl",
+        "//pkg/cloud/impl:cloudimpl",
         "//pkg/crosscluster/logical",
         "//pkg/crosscluster/physical",
         "//pkg/crosscluster/producer",
