--- pkg/ccl/cliccl/BUILD.bazel
@@ -1,58 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
-
-go_library(
-    name = "cliccl",
-    srcs = [
-        "cliccl.go",
-        "context.go",
-        "debug.go",
-        "demo.go",
-        "flags.go",
-        "mt.go",
-        "mt_http_test_directory.go",
-        "mt_proxy.go",
-        "mt_test_directory.go",
-    ],
-    importpath = "github.com/cockroachdb/cockroach/pkg/ccl/cliccl",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//pkg/ccl/securityccl/fipsccl",
-        "//pkg/ccl/sqlproxyccl",
-        "//pkg/ccl/sqlproxyccl/tenant",
-        "//pkg/ccl/sqlproxyccl/tenantdirsvr",
-        "//pkg/ccl/utilccl",
-        "//pkg/ccl/workloadccl/cliccl",
-        "//pkg/cli",
-        "//pkg/cli/clierror",
-        "//pkg/cli/clierrorplus",
-        "//pkg/cli/cliflagcfg",
-        "//pkg/cli/cliflags",
-        "//pkg/cli/democluster",
-        "//pkg/cli/exit",
-        "//pkg/roachpb",
-        "//pkg/util/log",
-        "//pkg/util/log/severity",
-        "//pkg/util/stop",
-        "//pkg/util/timeutil",
-        "@com_github_cockroachdb_errors//:errors",
-        "@com_github_cockroachdb_redact//:redact",
-        "@com_github_olekukonko_tablewriter//:tablewriter",
-        "@com_github_spf13_cobra//:cobra",
-        "@com_github_spf13_pflag//:pflag",
-        "@org_golang_google_grpc//:grpc",
-    ],
-)
-
-go_test(
-    name = "cliccl_test",
-    size = "medium",
-    srcs = ["main_test.go"],
-    data = glob(["testdata/**"]),
-    deps = [
-        "//pkg/base",
-        "//pkg/build",
-        "//pkg/ccl",
-        "//pkg/server",
-        "//pkg/testutils/serverutils",
-    ],
-)
