--- pkg/ccl/cliccl/cliccl.go
@@ -1,8 +1,0 @@
-// Copyright 2018 The Cockroach Authors.
-//
-// Use of this software is governed by the CockroachDB Software License
-// included in the /LICENSE file.
-
-package cliccl
-
-import _ "github.com/cockroachdb/cockroach/pkg/ccl/workloadccl/cliccl" // init hooks
