--- pkg/ccl/cliccl/main_test.go
@@ -1,30 +1,0 @@
-// Copyright 2017 The Cockroach Authors.
-//
-// Use of this software is governed by the CockroachDB Software License
-// included in the /LICENSE file.
-
-package cliccl_test
-
-import (
-	"os"
-	"testing"
-
-	"github.com/cockroachdb/cockroach/pkg/base"
-	"github.com/cockroachdb/cockroach/pkg/build"
-	"github.com/cockroachdb/cockroach/pkg/ccl"
-	"github.com/cockroachdb/cockroach/pkg/server"
-	"github.com/cockroachdb/cockroach/pkg/testutils/serverutils"
-)
-
-func TestMain(m *testing.M) {
-	// CLI tests are sensitive to the server version, but test binaries don't have
-	// a version injected. Pretend to be a very up-to-date version.
-	defer build.TestingOverrideVersion("v999.0.0")()
-
-	defer ccl.TestingEnableEnterprise()()
-	serverutils.InitTestServerFactory(server.TestServerFactory,
-		serverutils.WithDRPCOption(base.TestDRPCEnabledRandomly))
-	os.Exit(m.Run())
-}
-
-//go:generate ../../util/leaktest/add-leaktest.sh *_test.go
