+++ pkg/ccl/telemetryccl/BUILD.bazel
@@ -14,6 +14,7 @@ go_test(
         "//pkg/base",
         "//pkg/ccl",
         "//pkg/ccl/multiregionccl/multiregionccltestutils",
+        "//pkg/cloud/impl:cloudimpl",
         "//pkg/cloud/nodelocal",
         "//pkg/jobs",
         "//pkg/jobs/jobspb",
