+++ pkg/cli/BUILD.bazel
@@ -27,6 +27,7 @@ go_library(
         "debug.go",
         "debug_check_store.go",
         "debug_ear.go",
+        "debug_fips.go",
         "debug_job_cleanup.go",
         "debug_job_trace.go",
         "debug_list_files.go",
@@ -56,7 +57,11 @@ go_library(
         "log_flags.go",
         "mt.go",
         "mt_cert.go",
+        "mt_http_test_directory.go",
+        "mt_proxy.go",
+        "mt_proxy_context.go",
         "mt_start_sql.go",
+        "mt_test_directory.go",
         "node.go",
         "nodelocal.go",
         "prefixer.go",
@@ -120,6 +125,10 @@ go_library(
     deps = [
         "//pkg/base",
         "//pkg/build",
+        "//pkg/ccl/securityccl/fipsccl",
+        "//pkg/ccl/sqlproxyccl",
+        "//pkg/ccl/sqlproxyccl/tenant",
+        "//pkg/ccl/sqlproxyccl/tenantdirsvr",
         "//pkg/cli/clicfg",
         "//pkg/cli/clientflags",
         "//pkg/cli/clienturl",
@@ -290,6 +299,7 @@ go_library(
         "@com_github_marusama_semaphore//:semaphore",
         "@com_github_mattn_go_isatty//:go-isatty",
         "@com_github_mozillazg_go_slugify//:go-slugify",
+        "@com_github_olekukonko_tablewriter//:tablewriter",
         "@com_github_spf13_cobra//:cobra",
         "@com_github_spf13_cobra//doc",
         "@com_github_spf13_pflag//:pflag",
