+++ pkg/cmd/cockroach/main.go
@@ -11,8 +11,7 @@
 package main
 
 import (
-	_ "github.com/cockroachdb/cockroach/pkg/ccl"        // ccl init hooks
-	_ "github.com/cockroachdb/cockroach/pkg/ccl/cliccl" // cliccl init hooks
+	_ "github.com/cockroachdb/cockroach/pkg/ccl" // ccl init hooks
 	"github.com/cockroachdb/cockroach/pkg/cli"
 	_ "github.com/cockroachdb/cockroach/pkg/ui/distccl" // ccl web UI init hook
 )
